LayoutTests/imported/w3c/web-platform-tests/editing/other/editing-around-select-element.tentative_forwardDelete-expected.txt


PASS Test execCommand with selection around select element
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>ab[c</p><select><option>d]ef</option></select></div>: shouldn't modify in <option> assert_in_array: value "<div contenteditable><p>ab<select><option>def</option></select></p></div>" not in array ["<div contenteditable><p>abc</p><select><option>def</option></select></div>", "<div contenteditable><p>ab</p><select><option>def</option></select></div>"]
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select><option>d[]ef</option></select></div>: shouldn't modify in <option> assert_equals: expected "<div contenteditable><p>abc</p><select><option>def</option></select></div>" but got "<div contenteditable><p>abc</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><select><option>ab[c</option></select><p>d]ef</p></div>: shouldn't modify in <option> assert_in_array: value "<div contenteditable><select><option>abc</option></select>ef</div>" not in array ["<div contenteditable><select><option>abc</option></select><p>def</p></div>", "<div contenteditable><select><option>abc</option></select><p>ef</p></div>"]
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select><option>{}def</option></select><p>ghi</p></div>: shouldn't modify in <option> assert_equals: expected "<div contenteditable><p>abc</p><select><option>def</option></select><p>ghi</p></div>" but got "<div contenteditable><p>abc</p><p>ghi</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select><option>def{}</option></select><p>ghi</p></div>: shouldn't modify in <option> assert_equals: expected "<div contenteditable><p>abc</p><select><option>def</option></select><p>ghi</p></div>" but got "<div contenteditable><p>abc</p><p>ghi</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select><option>{def}</option></select><p>ghi</p></div>: shouldn't modify in <option> assert_equals: expected "<div contenteditable><p>abc</p><select><option>def</option></select><p>ghi</p></div>" but got "<div contenteditable><p>abc</p><p>ghi</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select><option>{def</option><option>ghi}</option></select><p>jkl</p></div>: shouldn't join <option>s assert_equals: expected "<div contenteditable><p>abc</p><select><option>def</option><option>ghi</option></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select>{<option>def</option>}<option>ghi</option></select><p>jkl</p></div>: shouldn't delete <option> assert_equals: expected "<div contenteditable><p>abc</p><select><option>def</option><option>ghi</option></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select><option>def</option>{<option>ghi</option>}</select><p>jkl</p></div>: shouldn't delete <option> assert_equals: expected "<div contenteditable><p>abc</p><select><option>def</option><option>ghi</option></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select>{<option>def</option><option>ghi</option>}</select><p>jkl</p></div>: shouldn't delete <option>s nor <select> assert_equals: expected "<div contenteditable><p>abc</p><select><option>def</option><option>ghi</option></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select><optgroup>{<option>def</option><option>ghi</option>}</optgroup></select><p>jkl</p></div>: shouldn't delete <option>, <optgroup> nor <select> assert_equals: expected "<div contenteditable><p>abc</p><select><optgroup><option>def</option><option>ghi</option></optgroup></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
PASS execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p>{<select><option>def</option><option>ghi</option></select>}<p>jkl</p></div>: <select> element itself should be removable
PASS execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p>{<select><optgroup><option>def</option><option>ghi</option></optgroup></select>}<p>jkl</p></div>: <select> element itself should be removable
PASS execCommand(forwardDelete, false, "") in <select contenteditable>{<option>abc</option><option>def</option>}</select>: shouldn't delete <option>s
PASS execCommand(forwardDelete, false, "") in <select><option contenteditable>{abc}</option><option>def</option></select>: shouldn't modify <option>
PASS execCommand(forwardDelete, false, "") in <select><optgroup contenteditable>{<option>abc</option><option>def</option>}</optgroup></select>: shouldn't delete <option>s
PASS execCommand(forwardDelete, false, "") in <select><optgroup contenteditable><option>{abc}</option><option>def</option></optgroup></select>: shouldn't delete <option>s nor optgroup
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>ab[c</p><select multiple><option>d]ef</option></select></div>: shouldn't modify in <option> assert_in_array: value "<div contenteditable><p>ab<select multiple><option>def</option></select></p></div>" not in array ["<div contenteditable><p>abc</p><select multiple><option>def</option></select></div>", "<div contenteditable><p>ab</p><select multiple><option>def</option></select></div>"]
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple><option>d[]ef</option></select></div>: shouldn't modify in <option> assert_equals: expected "<div contenteditable><p>abc</p><select multiple><option>def</option></select></div>" but got "<div contenteditable><p>abc</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><select multiple><option>ab[c</option></select><p>d]ef</p></div>: shouldn't modify in <option> assert_in_array: value "<div contenteditable><select multiple><option>abc</option></select>ef</div>" not in array ["<div contenteditable><select multiple><option>abc</option></select><p>def</p></div>", "<div contenteditable><select multiple><option>abc</option></select><p>ef</p></div>"]
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple><option>{}def</option></select><p>ghi</p></div>: shouldn't modify in <option> assert_equals: expected "<div contenteditable><p>abc</p><select multiple><option>def</option></select><p>ghi</p></div>" but got "<div contenteditable><p>abc</p><p>ghi</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple><option>def{}</option></select><p>ghi</p></div>: shouldn't modify in <option> assert_equals: expected "<div contenteditable><p>abc</p><select multiple><option>def</option></select><p>ghi</p></div>" but got "<div contenteditable><p>abc</p><p>ghi</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple><option>{def}</option></select><p>ghi</p></div>: shouldn't modify in <option> assert_equals: expected "<div contenteditable><p>abc</p><select multiple><option>def</option></select><p>ghi</p></div>" but got "<div contenteditable><p>abc</p><p>ghi</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple><option>{def</option><option>ghi}</option></select><p>jkl</p></div>: shouldn't join <option>s assert_equals: expected "<div contenteditable><p>abc</p><select multiple><option>def</option><option>ghi</option></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple>{<option>def</option>}<option>ghi</option></select><p>jkl</p></div>: shouldn't delete <option> assert_equals: expected "<div contenteditable><p>abc</p><select multiple><option>def</option><option>ghi</option></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple><option>def</option>{<option>ghi</option>}</select><p>jkl</p></div>: shouldn't delete <option> assert_equals: expected "<div contenteditable><p>abc</p><select multiple><option>def</option><option>ghi</option></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple>{<option>def</option><option>ghi</option>}</select><p>jkl</p></div>: shouldn't delete <option>s nor <select multiple> assert_equals: expected "<div contenteditable><p>abc</p><select multiple><option>def</option><option>ghi</option></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
FAIL execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p><select multiple><optgroup>{<option>def</option><option>ghi</option>}</optgroup></select><p>jkl</p></div>: shouldn't delete <option>, <optgroup> nor <select multiple> assert_equals: expected "<div contenteditable><p>abc</p><select multiple><optgroup><option>def</option><option>ghi</option></optgroup></select><p>jkl</p></div>" but got "<div contenteditable><p>abc</p><p>jkl</p></div>"
PASS execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p>{<select multiple><option>def</option><option>ghi</option></select>}<p>jkl</p></div>: <select multiple> element itself should be removable
PASS execCommand(forwardDelete, false, "") in <div contenteditable><p>abc</p>{<select multiple><optgroup><option>def</option><option>ghi</option></optgroup></select>}<p>jkl</p></div>: <select multiple> element itself should be removable
PASS execCommand(forwardDelete, false, "") in <select multiple contenteditable>{<option>abc</option><option>def</option>}</select>: shouldn't delete <option>s
PASS execCommand(forwardDelete, false, "") in <select multiple><option contenteditable>{abc}</option><option>def</option></select>: shouldn't modify <option>
PASS execCommand(forwardDelete, false, "") in <select multiple><optgroup contenteditable>{<option>abc</option><option>def</option>}</optgroup></select>: shouldn't delete <option>s
PASS execCommand(forwardDelete, false, "") in <select multiple><optgroup contenteditable><option>{abc}</option><option>def</option></optgroup></select>: shouldn't delete <option>s nor optgroup
PASS execCommand(forwardDelete, false, "") in <optgroup contenteditable><option>{abc}</option><option>def</option></optgroup>: shouldn't delete <option>s nor optgroup
PASS execCommand(forwardDelete, false, "") in <option contenteditable>{abc}</option>: shouldn't modify <option>
